requirements.txt
angr==9.2.119
cupy_cuda12x==13.3.0
fire==0.6.0
matplotlib==3.9.2
networkx==3.3
numpy==2.1.2
pandas==2.2.3
polars==1.9.0
pyarrow==17.0.0
scalib==0.6.0
scikit_learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
tqdm==4.66.5
